refactor utils to child process files
